refactor(dashboard): extract sub-components to components/dashboard/
DashboardPage.tsx had 16 presentational sub-components defined inline,
inflating the file to 1309 lines. Each panel is now independently
importable and testable.

- tokens.ts: shared C color aliases (extracted from DashboardPage)
- panels.tsx: all presentational components (Card, StatCards, Maintenance,
  Documents, Insights, ValueTracker, QuickActions, Activity, Neighbors, Properties)
- AIAssistantPanel.tsx: separated because it owns useVoiceAgent state
- DashboardPage.tsx: reduced from 1309 → 455 lines; hooks + derived data + render only

No logic changes; dead `jobs` prop removed from MyPropertiesPanel.
