+++ frontend/src/components/dashboard/tokens.ts
@@ -1,0 +1,21 @@
+import { COLORS, FONTS } from "@/theme";
+
+export { FONTS };
+
+export const C = {
+  bg:       COLORS.canvas,
+  card:     "#FFFFFF",
+  border:   "#E5E7EB",
+  text:     COLORS.plum,
+  muted:    COLORS.plumMid,
+  green:    COLORS.sageText,
+  greenBg:  COLORS.sageLight,
+  greenBdr: COLORS.sageMid,
+  blue:     "#2563EB",
+  blueBg:   "#EFF6FF",
+  orange:   "#D97706",
+  orangeBg: "#FFFBEB",
+  red:      COLORS.errorText,
+  redBg:    "#FEF2F2",
+  shadow:   "0 1px 3px rgba(0,0,0,0.06), 0 1px 2px rgba(0,0,0,0.04)",
+};
